+++ Procfile
@@ -1,1 +1,1 @@
-worker2: python pulse_actions/worker.py --config-file configs/production_config.json
+worker2: python pulse_actions/worker.py --config-file configs/production_config.json --dry-run False
