--- pulse_actions/run_time_config.json
@@ -1,31 +1,0 @@
-{
-    "manual_backfill": {
-        "exchange": "exchange/treeherder/v1/job-actions",
-        "topic": "buildbot.#.backfill"
-    },
-    "manual_backfill-stage": {
-        "exchange": "exchange/treeherder-stage/v1/job-actions",
-        "topic": "buildbot.#.backfill"
-    },
-    "resultset_actions": {
-        "exchange": "exchange/treeherder/v1/resultset-actions",
-        "topic": "#.#"
-    },
-    "resultset_actions-stage": {
-        "exchange": "exchange/treeherder-stage/v1/resultset-actions",
-        "topic": "#.#"
-    },
-    "backfilling": {
-        "exchange": "exchange/build/normalized",
-        "topic": "unittest.mozilla-inbound.#"
-    },
-    "runnable": {
-        "exchange": "exchange/treeherder/v1/resultset-runnable-job-actions",
-        "topic": "#"
-    },
-    "runnable-stage": {
-        "exchange": "exchange/treeherder-stage/v1/resultset-runnable-job-actions",
-        "topic": "#"
-    }
-
-}
